refactor: decompose large files by extracting utilities and helpers
Extract utilities and type definitions from oversized files to improve
maintainability and reduce file complexity:

- demo.ts (1,656→499 lines): Extract demo-uuid.ts, demo-compensation.ts,
  and demo-generators.ts for UUID generation, compensation calculations,
  and sample data generation
- client.ts (1,033→602 lines): Extract form-serialization.ts,
  error-handling.ts, and property-configs.ts for API utilities
- ValidateGameModal.tsx (698→541 lines): Extract UnsavedChangesDialog
  and ValidationSuccessToast components
- useValidationState.ts (573→240 lines): Extract validation/types.ts
  and validation/api-helpers.ts for type definitions and API helpers
- auth.ts (563→274 lines): Extract auth-parsers.ts for HTML parsing
  and login submission utilities

Fixes #249
