+++ web-app/src/components/features/validation/index.ts
@@ -6,3 +6,5 @@ export { ScoresheetPanel } from "./ScoresheetPanel";
 export { AddPlayerSheet } from "./AddPlayerSheet";
 export { RosterVerificationPanel } from "./RosterVerificationPanel";
 export { PlayerListItem } from "./PlayerListItem";
+export { UnsavedChangesDialog } from "./UnsavedChangesDialog";
+export { ValidationSuccessToast } from "./ValidationSuccessToast";
